refactor: migrate init command to use Pack wrapper pattern
- Add Initialize function to pkg/pack for creating new packs
- Move pack initialization logic from commands/initialize to pack package
- Update init command to delegate to pack.Initialize
- Move commentOutConfigValues helper to pack package
- Remove test for commentOutConfigValues from command tests

This continues the pattern established with the fill command, moving
pack-specific logic into the pack wrapper package.
