+++ pkg/commands/initialize/init_test.go
@@ -223,59 +223,3 @@ func TestInitPackIntegration(t *testing.T) {
 		assert.True(t, info.IsDir())
 	})
 }
-
-func TestCommentOutConfigValues(t *testing.T) {
-	tests := []struct {
-		name     string
-		input    string
-		expected string
-	}{
-		{
-			name: "comments out values but keeps structure",
-			input: `# This is a comment
-[pack]
-ignore = ["*.bak"]
-
-[symlink]
-force_home = ["ssh"]`,
-			expected: `# This is a comment
-[pack]
-# ignore = ["*.bak"]
-
-[symlink]
-# force_home = ["ssh"]`,
-		},
-		{
-			name: "preserves existing comments",
-			input: `# Already commented
-# ignore = ["test"]
-[section]
-value = "data"`,
-			expected: `# Already commented
-# ignore = ["test"]
-[section]
-# value = "data"`,
-		},
-		{
-			name:     "handles empty input",
-			input:    "",
-			expected: "",
-		},
-		{
-			name: "preserves blank lines",
-			input: `[section]
-
-value = "test"`,
-			expected: `[section]
-
-# value = "test"`,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := commentOutConfigValues(tt.input)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
